+++ docs/pages/components/message/Message.vue
@@ -16,6 +16,10 @@
 
         <Example :component="ExSizes" :code="ExSizesCode" title="Sizes" vertical/>
 
+        <Example :component="ExAutoClose" :code="ExAutoCloseCode" title="Auto Close" vertical>
+            <p>Message will be automatically closed after <code>duration</code>.</p>
+        </Example>
+
         <ApiView :data="api"/>
     </div>
 </template>
@@ -38,6 +42,9 @@
     import ExSizes from './examples/ExSizes'
     import ExSizesCode from '!!raw-loader!./examples/ExSizes'
 
+    import ExAutoClose from './examples/ExAutoClose'
+    import ExAutoCloseCode from '!!raw-loader!./examples/ExAutoClose'
+
     export default {
         data() {
             return {
@@ -47,11 +54,13 @@
                 ExIcons,
                 ExHeaderless,
                 ExSizes,
+                ExAutoClose,
                 ExSimpleCode,
                 ExTypesCode,
                 ExIconsCode,
                 ExHeaderlessCode,
-                ExSizesCode
+                ExSizesCode,
+                ExAutoCloseCode
             }
         }
     }